+++ pkg/R/modEvAmethods.R
@@ -1,8 +1,8 @@
 modEvAmethods <- function(fun) {
-  # version 1.6 (25 Mar 2024)
+  # version 1.7 (28 Apr 2026)
 
   if (fun %in% c("threshMeasures", "multModEv")) {
-    thresh.measures <- c("CCR", "Misclass", "Sensitivity", "Specificity", "Omission", "Commission", "Precision", "Recall", "PPP", "NPP", "UPR", "OPR", "PPI", "PAI", "kappa", "TSS", "NMI", "F1score", "OddsRatio", "ORSS", "SEDI")
+    thresh.measures <- c("CCR", "Misclass", "Sensitivity", "Specificity", "Omission", "Commission", "Precision", "Recall", "PPP", "NPP", "UPR", "OPR", "PPI", "PAI", "kappa", "TSS", "NMI", "F1score", "OddsRatio", "ORSS", "SEDI", "AVI", "CVI")
 
     if (fun == "threshMeasures") return(thresh.measures)